Leaf-ETL/requirements.txt
flask==3.0.3
google-cloud-firestore==2.17.0
google-auth==2.32.0
psycopg2-binary==2.9.9
pytest
pytest-html
flask_cors
pyzipper~=0.3.6